feat: move auth logic into useAuth
